+++ packages/base-data-service/tests/ExampleDataService.ts
@@ -1,6 +1,6 @@
-import { ConstantBackoff } from '@metamask/controller-utils';
 import { Messenger } from '@metamask/messenger';
 import { CaipAssetId, Duration, inMilliseconds, Json } from '@metamask/utils';
+import { ConstantBackoff } from 'cockatiel';
 
 import {
   BaseDataService,